feat(dashboard): RoundStatus and DashboardForm are now aware of ICO stages.
This reduces maintenance, since the dates are fixed on a smart contract
an thus unlikely to be changed. Therefore the code can reflect this fact
and reduce the need to change the values, which are known upfront.

The existing RoundStatus is refactored into RoundStage, receiving the ongoing
stage as `stage` prop.

RoundStatus becomes aware of current date and selects the ongoing stage
from a hardcoded list, managing a RoundStage.

The DashboardForm also gets its conversion rate from the current stage
(and passes it along to the calculator).